+++ client/src/components/SignupAndLogin/Signup/Signup.css
@@ -1,5 +1,5 @@
 .signupForm {
-    width: 65%;
+    width: 55%;
     margin: auto;
     margin-top: 100px;
 }
