+++ client/src/components/StaticComps/Footer/index.js
