+++ client/src/components/StaticComps/Map/TutorLogo.png
