+++ client/src/components/StaticComps/Map/index.js
